ITN fixes
An assortment of fixes following the last period to make the API
more stable. We also remove prematurer deduplication so that we can
now identify when validators are submitting too many values.

For the HTMX endpoint we also handle a case where location data
cannot be retrieved initially meaning we attempted to process
a dict when were were receiving a string.