Add price comparisons tables
This is very very rough code to try and extract information about
price deviations among participants. It may be better presented
visually than it is as a table. The data should be there to do that
but not the UI skills! Let's test this first and see how it goes.